+++ src/main/clojure/zi/codox.clj
@@ -4,6 +4,8 @@
    [zi.mojo :as mojo]
    [zi.core :as core]
    [clojure.java.io :as io])
+  (:use
+   [zi.maven :only [resolve-dependency]])
   (:import
    java.io.File
    [clojure.maven.annotations
@@ -11,37 +13,45 @@
    org.apache.maven.plugin.MojoExecutionException
    org.apache.maven.artifact.Artifact))
 
-(def ^{:const true} codox-path-regex
-  #"codox/codox/[0-9.]+(?:-SNAPSHOT)?/codox")
-
-(def ^{:const true} tools-namespace-path-regex
-  #"org/clojure/tools.namespace/[0-9.]+(?:-SNAPSHOT)?/tools.namespace")
-
-(def ^{:const true} java-classpath-path-regex
-  #"org/clojure/java.classpath/[0-9.]+(?:-SNAPSHOT)?/java.classpath")
-
-(def ^{:const true} hiccup-path-regex
-  #"hiccup/hiccup/[0-9.]+(?:-SNAPSHOT)?/hiccup")
-
 (defn run-codox
-  [project source-paths classpath-elements target-path version writer]
-  (core/eval-clojure
-   source-paths
-   (vec (concat classpath-elements (core/zi-classpath-elements)))
-   `(do
-      (require 'codox.main)
-      (codox.main/generate-docs
-       {:name ~(.getName project)
-        :version ~version
-        :description ~(.getDescription project)
-        :sources ~source-paths
-        :output-dir ~target-path
-        :writer '~writer}))))
+  [repo-system repo-system-session project
+   source-paths classpath-elements target-path version writer]
+  (let [codox-deps (resolve-dependency
+                    repo-system
+                    repo-system-session
+                    (.getRemoteProjectRepositories project)
+                    "codox"
+                    (or (System/getProperty "codox.artifact-id") "codox.core")
+                    (or (System/getProperty "codox.version") "0.6.1")
+                    {})
+        codox-md-deps (resolve-dependency
+                       repo-system
+                       repo-system-session
+                       (.getRemoteProjectRepositories project)
+                       "codox-md" "codox-md"
+                       (or (System/getProperty "codox-md.version") "0.1.0")
+                       {})]
+    (core/eval-clojure
+     source-paths
+     (vec (concat classpath-elements codox-deps codox-md-deps))
+     `(do
+        (require 'codox.main)
+        (codox.main/generate-docs
+         {:name ~(.getName project)
+          :version ~version
+          :description ~(core/unindent-description (.getDescription project))
+          :sources ~source-paths
+          :output-dir ~target-path
+          :writer '~writer})))))
 
 (mojo/defmojo Codox
   {Goal "codox"
    RequiresDependencyResolution "test"}
-  [^{Parameter
+  [^{Component {:role "org.sonatype.aether.RepositorySystem"}}
+   repoSystem
+   ^{Parameter {:defaultValue "${repositorySystemSession}" :readonly true}}
+   repoSystemSession
+   ^{Parameter
      {:defaultValue "${project.build.directory}/doc"
       :alias "codoxTargetDirectory"
       :description "Where to write codox output"}}
@@ -65,6 +75,8 @@
   (let [writer (symbol (or codox-writer "codox.writer.html/write-docs"))]
     (.debug log (str "Writer is " writer))
     (run-codox
+     repoSystem
+     repoSystemSession
      project
      (core/clojure-source-paths source-directory)
      test-classpath-elements