+++ src/main/clojure/zi/lein.clj
@@ -1,10 +1,13 @@
 (ns zi.lein
-  "Leiningen project functions")
+  "Leiningen project functions"
+  (:require
+   [leiningen.core.project :as project]))
 
 (defn paths-for-checkout
   "Return the source paths for a given pom file"
   [project-file]
-  (let [project (read-string (slurp project-file))]
-    [(:source-path project "src")
-     (:compile-path project "classes")
-     (:resources-path project "resources")]))
+  (let [project (project/read project-file)]
+    (concat
+     (:source-paths project ["src"])
+     [(:compile-path project ["classes"])]
+     (:resource-paths project ["resources"]))))